[vm] Re-do in terms of ported Contextify code.
Ported from https://github.com/brianmcd/contextify but with many
changes to fit into the Node.js codebase.

This is similar to node_script.cc, which formerly backed vm.js, but it
actually works well (see #3042).

As of now, it does not support the context <-> sandbox conflation that
the current vm module supports; the tests are updated to reflect that.
It also does not support the timeout parameter. Further commits will
reintroduce these features.

New tests are included to demonstrate what was fixed versus the current
vm module.