[vm] Reintroduce context <-> sandbox equality.
Contextify distinguishes between the sandbox, which is what is passed
in to the vm module and is the global object for a given context, and
the context itself, which is a C++ object in which scripts can run.

This commit hides the context from the user-facing interface, just like
in the vm module in Node.js 0.10 and lower. Users of the vm module only
deal with the sandbox objects, which internally are translated to their
corresponding contexts.